feat(ia): pr feedback - created individual modal files
added upsert, view and confirmation modals
